+++ packages/nimble-components/src/rich-text/mention-listbox/index.ts
@@ -54,6 +54,7 @@ export class RichTextMentionListbox extends FoundationListbox {
      *
      * @internal
      */
+    @observable
     public filteredOptions: ListboxOption[] = [];
 
     /**